Moves select gqls to use case hook
