+++ apps/judicial-system/web/src/routes/Court/Ruling/StepOne/RulingStepOne.tsx
@@ -47,7 +47,7 @@ import { isolation } from '@island.is/judicial-system-web/src/utils/Restrictions
 import CheckboxList from '@island.is/judicial-system-web/src/shared-components/CheckboxList/CheckboxList'
 import { useRouter } from 'next/router'
 import DateTime from '@island.is/judicial-system-web/src/shared-components/DateTime/DateTime'
-import useCase from '@island.is/judicial-system-web/src/utils/hooks/useCase'
+import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import { UserContext } from '@island.is/judicial-system-web/src/shared-components/UserProvider/UserProvider'
 
 export const RulingStepOne: React.FC = () => {
