+++ apps/judicial-system/web/src/routes/Prosecutor/InvestigationRequest/PoliceReport/PoliceReportForm.tsx
@@ -10,7 +10,7 @@ import {
   removeTabsValidateAndSet,
   validateAndSendToServer,
 } from '@island.is/judicial-system-web/src/utils/formHelper'
-import useCase from '@island.is/judicial-system-web/src/utils/hooks/useCase'
+import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import {
   FormSettings,
   useCaseFormHelper,
