+++ apps/judicial-system/web/src/routes/Prosecutor/RestrictionRequest/StepFour/StepFour.tsx
@@ -21,7 +21,7 @@ import {
   removeTabsValidateAndSet,
 } from '@island.is/judicial-system-web/src/utils/formHelper'
 import { useRouter } from 'next/router'
-import useCase from '@island.is/judicial-system-web/src/utils/hooks/useCase'
+import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import { formatProsecutorDemands } from '@island.is/judicial-system/formatters'
 
 export const StepFour: React.FC = () => {
