+++ apps/judicial-system/web/src/routes/Prosecutor/RestrictionRequest/StepThree/StepThreeForm.tsx
@@ -18,7 +18,7 @@ import {
   setCheckboxAndSendToServer,
   validateAndSendToServer,
 } from '@island.is/judicial-system-web/src/utils/formHelper'
-import useCase from '@island.is/judicial-system-web/src/utils/hooks/useCase'
+import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import CheckboxList from '@island.is/judicial-system-web/src/shared-components/CheckboxList/CheckboxList'
 import {
   custodyProvisions,
