+++ apps/judicial-system/web/src/routes/Prosecutor/RestrictionRequest/StepTwo/StepTwo.tsx
@@ -23,8 +23,10 @@ import * as Constants from '@island.is/judicial-system-web/src/utils/constants'
 import { CaseQuery } from '@island.is/judicial-system-web/graphql'
 import { UsersQuery } from '@island.is/judicial-system-web/src/utils/mutations'
 import { UserContext } from '@island.is/judicial-system-web/src/shared-components/UserProvider/UserProvider'
-import useCase from '@island.is/judicial-system-web/src/utils/hooks/useCase'
-import useInstitution from '@island.is/judicial-system-web/src/utils/hooks/useInstitution'
+import {
+  useCase,
+  useInstitution,
+} from '@island.is/judicial-system-web/src/utils/hooks'
 import StepTwoForm from './StepTwoForm'
 
 export const StepTwo: React.FC = () => {
