+++ apps/judicial-system/web/src/routes/Prosecutor/SharedComponents/DefendantInfo/DefendantInfo.tsx
@@ -8,7 +8,7 @@ import {
   setAndSendToServer,
   validateAndSendToServer,
 } from '@island.is/judicial-system-web/src/utils/formHelper'
-import useCase from '@island.is/judicial-system-web/src/utils/hooks/useCase'
+import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import * as styles from './DefendantInfo.treat'
 
 interface Props {
