+++ apps/judicial-system/web/src/routes/Prosecutor/SharedComponents/LokeCaseNumber/LokeCaseNumber.tsx
@@ -6,7 +6,7 @@ import {
   removeTabsValidateAndSet,
   validateAndSendToServer,
 } from '@island.is/judicial-system-web/src/utils/formHelper'
-import useCase from '@island.is/judicial-system-web/src/utils/hooks/useCase'
+import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 
 interface Props {
   workingCase: Case
