+++ apps/judicial-system/web/src/routes/Prosecutor/SharedComponents/SelectCourt/SelectCourt.tsx
@@ -4,7 +4,7 @@ import { ReactSelectOption } from '@island.is/judicial-system-web/src/types'
 import { setAndSendToServer } from '@island.is/judicial-system-web/src/utils/formHelper'
 import { Case, Institution } from '@island.is/judicial-system/types'
 import { ValueType } from 'react-select'
-import useCase from '@island.is/judicial-system-web/src/utils/hooks/useCase'
+import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 
 interface Props {
   workingCase: Case
