+++ apps/judicial-system/web/src/utils/hooks/use-case/createCourtCaseGql.ts
@@ -1,0 +1,9 @@
+import { gql } from '@apollo/client'
+
+export const CreateCourtCaseMutation = gql`
+  mutation CreateCourtCaseMutation($input: CreateCourtCaseInput!) {
+    createCourtCase(input: $input) {
+      courtCaseNumber
+    }
+  }
+`
