+++ apps/judicial-system/web/src/utils/hooks/use-case/updateCaseGql.ts
@@ -1,0 +1,10 @@
+import { gql } from '@apollo/client'
+
+export const UpdateCaseMutation = gql`
+  mutation UpdateCaseMutation($input: UpdateCaseInput!) {
+    updateCase(input: $input) {
+      id
+      modified
+    }
+  }
+`
