+++ apps/judicial-system/web/src/utils/useFormHelper.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import { Case } from '@island.is/judicial-system/types'
 import { parseString, replaceTabs } from './formatters'
 import { validate, Validation } from './validate'
-import useCase from './hooks/useCase'
+import { useCase } from './hooks'
 
 export interface FieldSettings {
   validations?: Validation[]
